# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Shift API
  description: |
    One API serving both the Shift web app and the future React Native app,
    with zero platform-specific branching in the API layer.

    * Auth uses Bearer access tokens (short-lived JWT) + rotating refresh tokens.
    * All mutating `POST`/`PATCH` requests accept an `Idempotency-Key` header.
    * All timestamps are ISO 8601 strings in UTC.
    * All IDs are UUIDv4 strings.
    * List endpoints that can grow unbounded use cursor-based pagination.
  version: "1.0.0"
  contact:
    name: Shift
servers:
  - url: https://api.shift.app/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /build-in-public:
    get:
      tags: [Build in Public]
      summary: List build-in-public posts
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated posts
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/BuildInPublicPost" }
    post:
      tags: [Build in Public]
      summary: Draft a build-in-public post and queue card image rendering
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [goalId, activityEntryId]
              properties:
                goalId: { type: string, format: uuid }
                activityEntryId: { type: string, format: uuid }
      responses:
        "201":
          description: Draft created; card image render queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BuildInPublicPost" }
  /build-in-public/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Build in Public]
      summary: Get a build-in-public post (includes cardImageUrl once ready)
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BuildInPublicPost" }
    patch:
      tags: [Build in Public]
      summary: Edit draft post text before posting
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                twitterText: { type: string }
                linkedinText: { type: string }
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BuildInPublicPost" }
  /build-in-public/{id}/render:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Build in Public]
      summary: (Re)render the card image for a post
      responses:
        "202":
          description: Render queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [queued] }
  /build-in-public/generate:
    post:
      tags: [Build in Public]
      summary: Generate multi-platform/multi-format content (threads, articles, summaries, ...) via the AI service
      description: |
        Separate from the activity-card flow above. Covers Twitter/LinkedIn
        short posts and threads, plus Medium/Dev.to/blog long-form content:
        founder updates, weekly summaries, milestone announcements,
        long-form articles, and technical blog posts. Content is generated
        by the AI service (NVIDIA primary, Anthropic fallback; see
        src/services/ai) and always fully editable afterward via
        PATCH /build-in-public/generated/{id}.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [goalId, platform, contentType]
              properties:
                goalId: { type: string, format: uuid }
                platform: { type: string, enum: [twitter, linkedin, medium, devto, blog] }
                contentType:
                  type: string
                  enum: [short_post, thread, founder_update, weekly_summary, milestone_announcement, long_form_article, technical_blog_post]
                activityEntryId:
                  type: string
                  format: uuid
                  description: Ties the post to one specific update. Omit to pull recent activity within timeframeDays instead.
                milestoneId:
                  type: string
                  format: uuid
                  description: For milestone_announcement. Defaults to the goal's most recently completed milestone if omitted.
                timeframeDays:
                  type: integer
                  description: How far back to pull activity for weekly_summary/founder_update/long-form formats. Default 7.
                tone:
                  type: string
                  description: Optional steering, e.g. "excited", "reflective".
      responses:
        "201":
          description: Generated content
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GeneratedPost" }
  /build-in-public/generated:
    get:
      tags: [Build in Public]
      summary: List generated content
      parameters:
        - name: goalId
          in: query
          schema: { type: string, format: uuid }
        - name: platform
          in: query
          schema: { type: string, enum: [twitter, linkedin, medium, devto, blog] }
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated generated content
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/GeneratedPost" }
  /build-in-public/generated/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Build in Public]
      summary: Get one generated content item
      responses:
        "200":
          description: Generated content
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GeneratedPost" }
    patch:
      tags: [Build in Public]
      summary: Edit generated content — fully editable, the model's output is a starting point
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string, nullable: true }
                segments:
                  type: array
                  items: { type: string }
                  description: One element for a single post; multiple, in order, for a thread.
      responses:
        "200":
          description: Updated content
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GeneratedPost" }
    delete:
      tags: [Build in Public]
      summary: Delete generated content
      responses:
        "204": { description: Deleted }

  # -------------------------------------------------------------- NOTIFICATIONS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived (15 min) access token. Refresh via POST /auth/refresh.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, format: uuid }
      description: Client-generated UUID. Server caches the response for 24h and replays it on retry.
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
      description: Opaque pagination cursor from a previous response's nextCursor.
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 30, minimum: 1, maximum: 100 }
...
  schemas:
...
    PaginatedResponse:
      type: object
      properties:
        items:
          type: array
          items: {}
        nextCursor:
          type: string
          nullable: true
...
    BuildInPublicPost:
      type: object
      properties:
        id: { type: string, format: uuid }
        goalId: { type: string, format: uuid }
        twitterText: { type: string }
        linkedinText: { type: string }
        cardHeadline: { type: string }
        cardSubline: { type: string }
        cardImageUrl: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    GeneratedPost:
      type: object
      properties:
        id: { type: string, format: uuid }
        goalId: { type: string, format: uuid }
        platform: { type: string, enum: [twitter, linkedin, medium, devto, blog] }
        contentType:
          type: string
          enum: [short_post, thread, founder_update, weekly_summary, milestone_announcement, long_form_article, technical_blog_post]
        title: { type: string, nullable: true }
        segments:
          type: array
          items: { type: string }
          description: One element for a single post; multiple, in order, for a thread.
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...